requirements.txt
psycopg2==2.9.3
python-decouple==3.6
aiogram==2.22.1
pyqiwip2p==2.0.6
